moved "loadRecentDocs" into Document.kt
